+++ packages/app/src/components/time-series-chart/logic/index.ts
@@ -5,5 +5,5 @@ export * from './common';
 export * from './legend';
 export * from './dimensions';
 export * from './curves';
-export * from './use-string-formatting';
+export * from './use-format-series-value';
 export * from './use-value-width';
